Introduced a conditional for mesh path inclusion
As of writing this, gazebo_ros cannot handle the package://-syntax
of mesh filenames. In order to work with gazebo, we'll have to rely
on absolute paths. This will however destroy visualizing the description
from a remote machine, as the absolute path there can (and most probably
will) be different.

This is considered a workaround until this issue is resolved upstream.