Remove rotation between base_link and kinematic chain
The robot's kinematic chain was rotated around the base_link, mainly
because of historic reasons. For this, an additional link called
`base_link_inertia` was introduced rotated 180 deg around base_link just
as `base`.

This commit removes this additional link which effectively rotates the
whole robot relative to base_link around 180 deg. This also makes
base_link go in line with base which should help reducing a lot of
confusion.